+++ src/main/java/de/telran/gartenshop/controller/CartController.java
@@ -1,7 +1,7 @@
 package de.telran.gartenshop.controller;
 
-import de.telran.gartenshop.dto.requestDto.CartItemRequestDto;
-import de.telran.gartenshop.dto.responseDto.CartItemResponseDto;
+import de.telran.gartenshop.dto.requestdto.CartItemRequestDto;
+import de.telran.gartenshop.dto.responsedto.CartItemResponseDto;
 import de.telran.gartenshop.service.CartService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
