+++ src/main/java/de/telran/gartenshop/controller/CategoryController.java
@@ -1,7 +1,7 @@
 package de.telran.gartenshop.controller;
 
-import de.telran.gartenshop.dto.requestDto.CategoryRequestDto;
-import de.telran.gartenshop.dto.responseDto.CategoryResponseDto;
+import de.telran.gartenshop.dto.requestdto.CategoryRequestDto;
+import de.telran.gartenshop.dto.responsedto.CategoryResponseDto;
 import de.telran.gartenshop.service.CategoryService;
 
 import lombok.RequiredArgsConstructor;
