+++ src/main/java/de/telran/gartenshop/dto/requestdto/FavoriteRequestDto.java
@@ -1,4 +1,4 @@
-package de.telran.gartenshop.dto.requestDto;
+package de.telran.gartenshop.dto.requestdto;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
@@ -9,9 +9,12 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class CartRequestDto {
+public class FavoriteRequestDto {
+    @NotNull(message = "ProductId cannot be null")
+    @Min(value = 1, message = "Invalid productId: productId must be >= 1")
+    private Long productId;
+
     @NotNull(message = "UserId cannot be null")
     @Min(value = 1, message = "Invalid userId: userId must be >= 1")
     private Long userId;
-
 }
