+++ src/main/java/de/telran/gartenshop/dto/responsedto/CartResponseDto.java
@@ -1,17 +1,15 @@
-package de.telran.gartenshop.dto.queryDto;
+package de.telran.gartenshop.dto.responsedto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class ProductProfitDto {
-    private String period;
-    private BigDecimal profit;
+public class CartResponseDto {
+    private Long cartId;
+    private Long userId;
 }
