+++ src/main/java/de/telran/gartenshop/security/jwt/JwtProvider.java
@@ -1,6 +1,6 @@
 package de.telran.gartenshop.security.jwt;
 
-import de.telran.gartenshop.dto.responseDto.UserResponseDto;
+import de.telran.gartenshop.dto.responsedto.UserResponseDto;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
