+++ src/test/java/de/telran/gartenshop/integration/CartIntegrationTest.java
@@ -1,13 +1,9 @@
 package de.telran.gartenshop.integration;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import de.telran.gartenshop.dto.requestDto.CartItemRequestDto;
-import de.telran.gartenshop.dto.requestDto.CartRequestDto;
-import de.telran.gartenshop.dto.requestDto.OrderRequestDto;
-import de.telran.gartenshop.dto.responseDto.*;
+import de.telran.gartenshop.dto.requestdto.CartItemRequestDto;
+import de.telran.gartenshop.dto.responsedto.*;
 import de.telran.gartenshop.entity.*;
-import de.telran.gartenshop.entity.enums.DeliveryMethod;
-import de.telran.gartenshop.entity.enums.OrderStatus;
 import de.telran.gartenshop.entity.enums.Role;
 import de.telran.gartenshop.repository.*;
 import de.telran.gartenshop.security.configure.SecurityConfig;
