+++ src/test/java/de/telran/gartenshop/integration/OrderIntegrationTest.java
@@ -1,8 +1,8 @@
 package de.telran.gartenshop.integration;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import de.telran.gartenshop.dto.requestDto.OrderRequestDto;
-import de.telran.gartenshop.dto.responseDto.*;
+import de.telran.gartenshop.dto.requestdto.OrderRequestDto;
+import de.telran.gartenshop.dto.responsedto.*;
 import de.telran.gartenshop.entity.*;
 import de.telran.gartenshop.entity.enums.DeliveryMethod;
 import de.telran.gartenshop.entity.enums.OrderStatus;
